refactor: calcTimeとcalculateTargetValuesのインポート元をutilsからlibに変更
